fix: AdminControllerTest의 오류를 수정한다
- restController와 일반 controller 분리 후 import를 제대로 해주지 않아서 오류 발생
